pyproject.toml
[tool.poetry]
name = "haystack-file-system"
version = "0.1.0"
description = "A distributed file system inspired by Haystack."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
redis = "^4.3.4"
psycopg2-binary = "^2.9.3"
requests = "^2.28.1"
pika = "^1.2.1"
faiss-cpu = "^1.7.2"
numpy = "^1.22.4"
aiofiles = "^0.8.0"
python-logging-loki = "^0.3.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
